+++ src/WayPoint/WayPointFile.cpp
@@ -149,7 +149,7 @@ WayPointFile::Parse(Waypoints &way_points,
     if (reader.error())
       return false;
 
-    double filesize = std::max(reader.size(), 1l);
+    unsigned filesize = std::max(reader.size(), 1l);
 
     // Read through the lines of the file
     TCHAR *line;
@@ -167,7 +167,7 @@ WayPointFile::Parse(Waypoints &way_points,
     if (reader.error())
       return false;
 
-    double filesize = std::max(reader.size(), 1l);
+    unsigned filesize = std::max(reader.size(), 1l);
 
     // Read through the lines of the file
     TCHAR *line;
